+++ lib/features/settings/components/rows/notification_time.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:pilll/features/localizations/l.dart';
 import 'package:pilll/utils/analytics.dart';
 import 'package:pilll/components/atoms/font.dart';
-import 'package:pilll/features/settings/reminder_times_page_legacy.dart';
+import 'package:pilll/features/reminder_times/page.dart';
 import 'package:pilll/entity/setting.codegen.dart';
 import 'package:pilll/utils/formatter/date_time_formatter.dart';
