+++ test/features/settings/reminder_times_widget_test.dart
@@ -2,7 +2,8 @@ import 'package:pilll/features/localizations/l.dart';
 import 'package:pilll/features/settings/provider.dart';
 import 'package:pilll/entity/pill_sheet_type.dart';
 import 'package:pilll/entity/setting.codegen.dart';
-import 'package:pilll/features/settings/reminder_times_page.dart';
+import 'package:pilll/features/settings/page.dart';
+import 'package:pilll/features/reminder_times/page.dart';
 import 'package:pilll/provider/setting.dart';
 import 'package:pilll/utils/environment.dart';
 import 'package:flutter/material.dart';
